services/langchain-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.1.20
langchain-openai==0.1.6
langchain-core==0.1.52
python-dotenv==1.0.0
python-multipart==0.0.6
openai>=1.24.0,<2.0.0

